consolidating encode/decode and property type functions into a separate file
