+++ amqtt/mqtt5/__init__.py
@@ -1,0 +1,1 @@
+"""INIT."""
