+++ amqtt/mqtt5/protocol/__init__.py
@@ -1,0 +1,1 @@
+"""INIT."""
